UI Clean up, Fixed selected image bug
When u get more info out of a movie, currently watching button should not be there.  There's no point to "currently watch" a movie. Just plan to watch it or complete it.

Whenever you would go from tv to movie the selected image would wiggle at the home button and then proceed. This was caused because the button and didscroll methods were connected. The solution I came up with was just adding a gesture recognizer for swiping left and right.